sudoku_env/server/requirements.txt
openenv-core[core]>=0.2.3
pydantic>=2.0
uvicorn>=0.24.0

